Add current directory to the JAVA_LIBRARY_PATH
When maven builds the agent jar it includes the libsigar-x86-linux.so into the jar, but
libsigar-x86-linux.so can't be loaded from the jar.
The alternative is to have the jar and libsigar-x86-linux.so separately and add the current 
directory to the JAVA_LIBRARY_PATH. This will allow libsigar-x86-linux.so to be loaded.